% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DocumentationConstructors.R
\name{discphasetype}
\alias{discphasetype}
\alias{contphasetype}
\title{Phase-type distributed objects}
\usage{
discphasetype(initDist, P.mat)

contphasetype(initDist, T.mat)
}
\arguments{
\item{initDist}{a vector holding the initial distribution of the phase-type
distribution. Hence \code{initDist} should satisfy that all entries are non-negative
and the sum of the vector must be less than or equal to 1.}

\item{P.mat}{A subtransition probability matrix satisfying that all entries
are non-negative. Furthermore, \code{P.mat} should be a square matrix, which
determinant is different from zero. All rows in \code{P.mat} should sum to a
number less than or equal to one, and the length of the initial distribution has to be
equal to the number of rows of \code{P.mat}.}

\item{T.mat}{A subintensity rate matrix. The matrix must be invertible, all diagonal
entries must be negative and all non-diagonal entries must be non-negative. Also for each
row the sum of the entries must be non-positive.
Furthermore, the length of the initial distribution has to be
equal to the number of rows and number of columns of \code{T.mat}.}
}
\value{
\code{discphasetype} returns a list of class \emph{"discphasetype"}, while
\code{contcphasetype} returns a list of class \emph{"contphasetype"}. Both lists
hold
\itemize{
  \item \code{initDist} : The inital distribution
  \item \code{P.mat}/\code{T.mat} : The subtransition/subintensity matrix
}
}
\description{
Constructing phase-type distributions
}
\details{
\code{discphasetype} creates an object of class \emph{"discphasetype"}, i.e. the
object represents a discrete phase-type distribution.
\code{contphasetype}creates an object of class \emph{"contphasetype"}, i.e. the
object represents a continuous phase-type distribution.
}
\section{Functions}{
\itemize{
\item \code{discphasetype}: Creating an object of type discphasetype

\item \code{contphasetype}: Creating an object of type contphasetype
}}

\examples{
## For n=4, the time to the most recent common ancestor is
## phase-type distributed with
## initital distribution
initDist <- c(1,0,0)
## and sub-intensity rate matrix
T.mat <- matrix(c(-6,6,0,
                   0,-3,3,
                   0,0,-1), nrow = 3, ncol = 3, byrow = TRUE)

TMRCA <- contphasetype(initDist, T.mat)

## For theta=2, the number of segregating sites plus one is
## discrete phase-type distributed with
## initital distribution
initDist <- c(1,0,0,0)
## and sub-transition probability matrix
P.mat <- matrix(c(0.4, 0.3, 4/30, 2/30,
                   0, 0.5, 2/9, 1/9,
                   0, 0, 2/3, 0,
                   0, 0, 0, 2/3), nrow = 4, ncol = 4, byrow = TRUE)

S_Total <- discphasetype(initDist, P.mat)

}
