% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DocumentationNumberOfSegregatingSites.R
\name{dSegregatingSites}
\alias{dSegregatingSites}
\title{The number of segregating sites}
\source{
Mogens Bladt and Bo Friis Nielsen (2017):
\emph{ Matrix-Exponential Distributions in Applied Probability}.
Probability Theory and Stochastic Modelling (Springer), Volume 81.

Asger Hobolth, Arno Siri-Jégousse, Mogens Bladt (2019):
\emph{Phase-type distributions in population genetics}.
Theoretical Population Biology, 127, pp. 16-32.
}
\usage{
dSegregatingSites(n, theta, k, plot = FALSE)
}
\arguments{
\item{n}{the sample size (n >= 1).}

\item{theta}{the mutation parameter (theta > 0).}

\item{k}{a non-negative number or a non-negative vector.}

\item{plot}{a logical value indicating whether the function should
plot the density of the total number of segregating sites for the
given values of k.}
}
\value{
The function returns the probabilities \eqn{P(S=k)} for all values
of \eqn{k}. Hence, the returned object is of the same length as \eqn{k}.
If \code{plot=TRUE}, the function also plots the densities as a function of
\eqn{k}.
}
\description{
The density function of the total number of segregating sites
}
\details{
The density of the total number of segregating sites can be obtained
by the aid of the block counting process together with the reward
transformation and the discretization. For more information on this topic see \code{vignette("PhaseTypeGenetics")} or
Hobolth et al. (2019): \emph{Phase-type distributions in population genetics}.
}
\examples{

## Computing the density for a sample of n=5
dSegregatingSites(n=5,theta=2,k=5)
dSegregatingSites(n=5,theta=2,k=1:20, plot=TRUE)

## We apply the function for different sample sizes
## and theta=2
k_vec <- 0:15
theta <- 2
## Defining a matrix of results
Res_Mat <- dSegregatingSites(n = 1, theta = theta, k = k_vec)
## And Applying the function for all n in {2,...,20}
for(n in 2:20){

Res_Mat <- cbind(Res_Mat, dSegregatingSites(n = n, theta = theta, k = k_vec))
}

## We reproduce Figure 4.1 in John Wakeley (2009):
## "Coalescent Theory: An Introduction",
## Roberts and Company Publishers, Colorado.
## by using the package plot3D.
plot3D::hist3D(x=k_vec, y=1:20, z=Res_Mat, col = "grey", border = "black",
       xlab = "k", ylab = "n", zlab = "P(S=k)",
       main = "The probability function of the number of segregating sites",
       sub = expression(paste("The mutation parameter is ", theta,"= 2")),
       cex.main = 0.9, colkey = FALSE, zlim = c(0,0.4))

}
\seealso{
\code{\link{SiteFrequencies}}, \code{\link{dphasetype}}.
}
