% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DocumentationSum.R
\name{sum.discphasetype}
\alias{sum.discphasetype}
\alias{sum.contphasetype}
\title{Sum of phase-type distributions}
\source{
Mogens Bladt and Bo Friis Nielsen (2017):
\emph{ Matrix-Exponential Distributions in Applied Probability}.
Probability Theory and Stochastic Modelling (Springer), Volume 81.
}
\usage{
\method{sum}{discphasetype}(object1, object2)

\method{sum}{contphasetype}(object1, object2)
}
\arguments{
\item{object1, object2}{two objects of class \code{discphasetype} 
or \code{contphasetype} for which the sum should be computed.}
}
\value{
The function \code{sum} returns an object of type \code{discphasetype} 
or \code{contphasetype} (depending on the input) holding the ditribution 
of the sum of the input objects.
}
\description{
The sum of two independent discrete or continuous 
phase-type distributed variables with initial distributions
\code{initDist1} and \code{initDist2} as well as subtransition/subintensity 
matrices equal to \code{P.mat1}/\code{T.mat1} and  \code{P.mat2}/\code{T.mat2}.
}
\details{
In the discrete case, the sum of two phase-type distributed 
variables \eqn{tau1 ~ DPH_p(\alpha,S)} and \eqn{tau2 ~ DPH_q(\beta,T)} is 
again discrete phase-type distributed in the following way
\deqn{tau1 + tau2 ~ DPH_{p+q}((\alpha,0),cbind((S, s \beta),(0,T)) )}.
In the continous case, the sum of two phase-type distributed 
variables \eqn{X ~ PH_p(\alpha,S)} and \eqn{Y ~ PH_q(\beta,T)} is 
again continuous and phase-type distributed in the following way
\deqn{X + Y ~ PH_{p+q}((\alpha,0),cbind((S, s \beta),(0,T)) ).}
}
\section{Methods (by class)}{
\itemize{
\item \code{discphasetype}: computing the sum of two discrete phase-type distributions

\item \code{contphasetype}: computing the sum of two continuous phase-type distributions
}}

\examples{


}
\seealso{
\code{\link{sum}}.
}
