% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DocumentationSummaries.R
\name{summary.discphasetype}
\alias{summary.discphasetype}
\alias{summary.contphasetype}
\title{Summaries of phase-type distributed objects}
\usage{
\method{summary}{discphasetype}(object)

\method{summary}{contphasetype}(object)
}
\arguments{
\item{object}{an object for which a summary is desired. To be able to use
the summaries for phase-type distributed objects,the object has to be of
class \code{discphasetype} or \code{contphasetype}.}
}
\value{
\code{summary} prints
the initial distribution and the subtransition/subintensity matrix of the
phase-type distribution. If the sum of all entries in the initial distribution
is less than one, it also prints the defect size \code{(1-sum(initDist))}.
}
\description{
This \code{summary} function is a method that depends on the \code{\link{class}} of the argument.
For objects of class \code{discphasetype} or \code{contphasetype},
\code{summary} prints the inital distribution and the
subtransition/subintensity matrix of the phase-type distribution.
}
\examples{

summary(T_Total$n10)

}
\seealso{
\code{\link{summary}}, \code{\link{summary.glm}},
\code{\link{summary.lm}}.
}
