% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DocumentationDatasetT-Total.R
\docType{data}
\name{T_Total}
\alias{T_Total}
\title{The total branch length for a sample size of n = 5,10,20,50 and 100.}
\format{A list containing 5 objects of type \code{contphasetype}.}
\usage{
T_Total
}
\description{
A dataset containing the initial distributions and subintensity rate matrices
for the total branch length  (T_Total) 
for a sample size of n in {5,10,20,50,100}.
}
\keyword{datasets}
